+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/PageEditor.js
@@ -44,7 +44,7 @@ import DragPreview from './DragPreview';
 import {EditableDecorationProvider} from './fragment-content/EditableDecorationContext';
 import {EditableProcessorContextProvider} from './fragment-content/EditableProcessorContext';
 import {
-	CollectionItem,
+	CollectionItemWithControls,
 	CollectionWithControls,
 	ColumnWithControls,
 	ContainerWithControls,
@@ -57,7 +57,7 @@ import {DragDropManager} from './useDragAndDrop';
 
 const LAYOUT_DATA_ITEMS = {
 	[LAYOUT_DATA_ITEM_TYPES.collection]: CollectionWithControls,
-	[LAYOUT_DATA_ITEM_TYPES.collectionItem]: CollectionItem,
+	[LAYOUT_DATA_ITEM_TYPES.collectionItem]: CollectionItemWithControls,
 	[LAYOUT_DATA_ITEM_TYPES.column]: ColumnWithControls,
 	[LAYOUT_DATA_ITEM_TYPES.container]: ContainerWithControls,
 	[LAYOUT_DATA_ITEM_TYPES.dropZone]: DropZoneWithControls,
