+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/index.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-export {default as CollectionItem} from './CollectionItem';
+export {default as CollectionItemWithControls} from './CollectionItemWithControls';
 export {default as CollectionWithControls} from './CollectionWithControls';
 export {default as Column} from './Column';
 export {default as ColumnWithControls} from './ColumnWithControls';
